requirements.txt
streamlit==1.31.0
streamlit-authenticator==0.2.2
pyyaml==6.0 